requirements.txt
bs4
pandas
requests
openai
pyyaml
flake8
pytest
pytest-cov